select-sounds-program/requirements.txt
pymongo
mongoengine
tqdm
colorama
python-dateutil
bs4
dnspython